Refactor ID types from u64 to usize for better compatibility
Updated ClientId and RequestId type aliases in types.rs and adjusted related code in globals.rs, client.rs, and common.rs to reflect this change. This improves memory usage and aligns with the intended use of these identifiers.

+++ src/types.rs
@@ -5,8 +5,8 @@ use reqwest::{
 use std::sync::{Arc, RwLock};
 
 /// Type aliases for IDs
-pub type ClientId = u64;
-pub type RequestId = u64;
+pub type ClientId = usize;
+pub type RequestId = usize;
 pub type ProgressInfo = Arc<RwLock<RequestProgress>>;
 
 /// Wrapper for HeaderMap to make it FFI-compatible
